// types/guestProfileMaster.ts
export interface GuestProfilePayload {
  profileId: number;
  hotelId: number;
  title: string;
  guestName: string;
  dob: string;
  address: string;
  city: string;
  zipCode: string;
  country: string;
  nationality: string;
  ppNo: string;
  phone: string;
  email: string;
  createdOn: string;
  createdBy: string;
  updatedOn: string;
  updatedBy: string;
}